Add CodeBuild webhook
This will initiate a CodeBuild run whenever a pull request is opened
against the application or manifest repository. This will let developers
know if their changes somehow broke the Docker build or manifest
generation. It will also pre-build Docker images so that they're
available when the pull request is merged.